wit/{iterate,ordered}: move from x/ dir
